- add running tests to the CI pipeline - small refactor of shared type
