+++ tests/index.test.ts
@@ -56,6 +56,8 @@ describe("basic tests", () => {
     coerce.bigintCoerce,
   ]);
 
+  // TODO: split into *.test.ts files and add more tests
+
   test("smoke tests", () => {
     expect(() => c.number().parse(42)).not.toThrow();
     expect(() => c.number().parse("hello")).toThrow();
@@ -113,70 +115,6 @@ describe("basic tests", () => {
     const schema3 = c.string().email().array().max(4).nonEmpty();
     type InferredSchema3 = Infer<typeof schema3>;
     type _typeTest3 = Expect<Equal<InferredSchema3, [string, ...string[]]>>;
-  });
-
-  test("arrays", () => {
-    expect(() => c.array(c.string().min(2)).parse(["ab", "ba"])).not.toThrow();
-    expect(() => c.array(c.string()).parse([])).not.toThrow();
-    expect(() => c.array(c.string()).parse([42])).toThrow();
-    expect(() =>
-      c.array(c.string()).nonEmpty().parse(["a", "b"]),
-    ).not.toThrow();
-    expect(() => c.array(c.string()).nonEmpty().parse([])).toThrow();
-    expect(() =>
-      c
-        .array(c.string())
-        .nonEmpty()
-        .min(3)
-        .max(5)
-        .length(4)
-        .parse(["a", "b", "c", "d"]),
-    ).not.toThrow();
-    expect(() =>
-      c
-        .array(c.string())
-        .nonEmpty()
-        .min(3)
-        .max(5)
-        .length(4)
-        .parse(["a", "b", "d"]),
-    ).toThrow();
-
-    const schema = c.array(c.string());
-    type SchemaType = Infer<typeof schema>;
-    type _test = Expect<Equal<SchemaType, string[]>>;
-  });
-
-  test("arrays alternative syntax", () => {
-    expect(() => c.string().min(2).array().parse(["ab", "ba"])).not.toThrow();
-    expect(() => c.string().array().parse([])).not.toThrow();
-    expect(() => c.string().array().parse([42])).toThrow();
-    expect(() => c.string().array().nonEmpty().parse(["a", "b"])).not.toThrow();
-    expect(() => c.string().array().nonEmpty().parse([])).toThrow();
-    expect(() =>
-      c
-        .string()
-        .array()
-        .nonEmpty()
-        .min(3)
-        .max(5)
-        .length(4)
-        .parse(["a", "b", "c", "d"]),
-    ).not.toThrow();
-    expect(() =>
-      c
-        .string()
-        .array()
-        .nonEmpty()
-        .min(3)
-        .max(5)
-        .length(4)
-        .parse(["a", "b", "d"]),
-    ).toThrow();
-
-    const schema = c.string().array();
-    type SchemaType = Infer<typeof schema>;
-    type _test = Expect<Equal<SchemaType, string[]>>;
   });
 
   test("sets", () => {
